initial_load changes are now consistent when in recursive Node's init
Node's init (and subsequent config file read) can be called recursively
in complex models (like debian dpkg). The previous way of starting
initial load at the beginning of init and stopping at the end did
not work in these recursive calls. Traces showed:

 1 on - 2 on - 3 off - 4 off

Hence, items loaded between step 3 and 4 were loaded out of initial mode,
which was wrong.

Now the recursitity is taken into account and the traces are now:

 1 on - 2 on - 3 on - 4 off

This change requires the default value of initial_load to be 0 (in
Instance. this was commited previously). Hence the changes to the
tests.