Make concurrent calls to madison to reduce user wait
Debian dependency checks needs to call madison system through http
to get a package status. Each call lasts for 2 or 3 seconds.

Before Memoize was used to cache the results, but each call was done
one by one. Which was not fun when a package feature 50 dependencies.

Now all calls are done in parrallel thanks to AnyEvent::HTTP. All
requests are sent and the dpkg backend will wait for all call to
finish before proceeding with a recheck of the dependencies (which go
faster since all data is cached).

madison results are still cached, but only with DB_File. Hence the
memoize dependency is removed.